fix: 총 수익률 계산 오류가 발생해 타입을 int에서 long으로 변경

+++ src/main/java/lotto/domain/Rank.java
@@ -13,9 +13,9 @@ public enum Rank {
     private static final int MATCHED_FOURTH_OR_THIRD = 5;
 
     private final int matched;
-    private final int prize;
+    private final long prize;
 
-    Rank(final int matched, final int prize) {
+    Rank(final int matched, final long prize) {
         this.matched = matched;
         this.prize = prize;
     }
@@ -37,7 +37,7 @@ public enum Rank {
         return matched;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
 }
